Renamed module entities
