+++ ml_project/setup.py
@@ -5,6 +5,6 @@ with open('requirements.txt') as f:
 
 setup(
     name="ml_project",
-    packages=['ml_project', 'tests', 'ml_project.enities', 'ml_project.models', 'ml_project.features'],
+    packages=['ml_project', 'tests', 'ml_project.entities', 'ml_project.models', 'ml_project.features'],
     install_requires=required,
 )
